+++ build-appimage.sh
@@ -9,31 +9,29 @@ mkdir -p "$APPDIR/usr/share/looksyk/static"
 mkdir -p "$APPDIR/usr/share/icons/hicolor/256x256/apps"
 mkdir -p "$APPDIR/usr/share/licenses/looksyk"
 
-# Backend-Binary kopieren
 cp target/looksyk "$APPDIR/usr/bin/looksyk-backend"
-# Electron-Client kopieren
-cp application-wrapper/looksyk "$APPDIR/usr/bin/looksyk"
-chmod +x "$APPDIR/usr/bin/looksyk"
+cp application-wrapper/Looksyk-AppRun "$APPDIR/AppRun"
 
-# AppRun ausführbar machen, falls vorhanden
 if [ -f "$APPDIR/AppRun" ]; then
   chmod +x "$APPDIR/AppRun"
 fi
 
-# Statische Dateien kopieren
 cp -r target/static/* "$APPDIR/usr/share/looksyk/static/"
 
-# app.asar kopieren
-cp target/application-wrapper/resources/app.asar "$APPDIR/usr/share/looksyk/app.asar"
+cp -r target/application-wrapper/* "$APPDIR/usr/share/looksyk/"
 
-# Icon kopieren
 cp icon/Looksyk_256.png "$APPDIR/usr/share/icons/hicolor/256x256/apps/de.sebastianruziczka.looksyk.png"
 
-# Desktop Datei kopieren
 cp application-wrapper/Looksyk.desktop "$APPDIR/looksyk.desktop"
 
-# Lizenz kopieren
 cp LICENSE "$APPDIR/usr/share/licenses/looksyk/LICENSE"
+
+if [ ! -f git-static ]; then
+  wget -O git-static https://www.kernel.org/pub/software/scm/git/git-2.9.5.tar.gz
+  tar -xzf git-static -C .
+  cp git-2.45.1-linux-x86_64-static/bin/git "$APPDIR/usr/bin/git"
+  rm -rf git-2.45.1-linux-x86_64-static git-static
+fi
 
 # AppImage bauen
 if [ ! -f linuxdeploy ]; then